Add "runBacktround" option in PoseNet
