--- Assets/Samples/PoseNet/PoseNetMuitlThreadSample.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b4c239042a38d493e9949d0106b19f74
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
