--- Assets/Samples/PoseNet/PoseNetMultiThread.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 015dc0c45125247c1a02a289b6d8adf9
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
