--- Assets/Samples/PoseNet/PoseNetTask.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c115bd63db5ef4dd8889f3de3f0344e6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
